# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Telecom System Api",
    "version": "1.0.0",
    "title": "API",
    "contact": {
      "email": "petersantoso94@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "schemes": ["http"],
  "host": "localhost:3000",
  "basePath": "/api",
  "paths": {
    "/users": {
      "post": {
        "summary": "Register user",
        "description": "Register user",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/createUserResponse"
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/sessions": {
      "get": {
        "summary": "Get logged in user",
        "description": "Get logged in user",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/userId"
            }
          }
        },
        "500": {
          "description": "Internal server error"
        }
      },
      "post": {
        "summary": "Login user",
        "description": "Login user",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation"
          }
        },
        "403": {
          "description": "Invalid status value"
        },
        "500": {
          "description": "Internal server error"
        }
      },
      "delete": {
        "summary": "Logout user",
        "description": "Logout user",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        },
        "500": {
          "description": "Internal server error"
        }
      }
    },
    "/inventory": {
      "post": {
        "summary": "Shipin inventory",
        "description": "Shipin inventory",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "shipin inventory request",
            "required": true,
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/shipinInventoryRequest" }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation"
          },
          "401": {
            "description": "Unautorized request"
          },
          "400": {
            "description": "Invalid request options",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/inventory/return": {
      "post": {
        "summary": "Return inventory",
        "description": "Return inventory",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Return inventory request",
            "required": true,
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/CommonInventoryRequest" }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation"
          },
          "401": {
            "description": "Unautorized request"
          },
          "400": {
            "description": "Invalid request options",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/inventory/shipout": {
      "post": {
        "summary": "Shipout inventory",
        "description": "Shipout inventory",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Shipout inventory request",
            "required": true,
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/CommonInventoryRequest" }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation"
          },
          "401": {
            "description": "Unautorized request"
          },
          "400": {
            "description": "Invalid request options",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/inventory/warehousemovement": {
      "post": {
        "summary": "Inventory Warehouse movement",
        "description": "Inventory Warehouse movement",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Inventory Warehouse movement request",
            "required": true,
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/CommonInventoryRequest" }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation"
          },
          "401": {
            "description": "Unautorized request"
          },
          "400": {
            "description": "Invalid request options",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "user": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "userId": {
      "type": "string"
    },
    "createUserResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "shipinInventoryRequest": {
      "type": "object",
      "properties": {
        "Serial_Number": {
          "type": "string"
        },
        "Price": {
          "type": "string"
        },
        "Missing": {
          "type": "boolean"
        },
        "Status": {
          "type": "string"
        },
        "Warehouse": {
          "type": "string"
        },
        "Type": {
          "type": "string"
        },
        "MSISDN": {
          "type": "string"
        },
        "Provider": {
          "type": "string"
        },
        "Voucher_Value": {
          "type": "string"
        },
        "Remark_Inv": {
          "type": "string"
        },
        "Fabiao_Number": {
          "type": "string"
        },
        "Form_Number": {
          "type": "string"
        }
      }
    },
    "CommonInventoryRequest": {
      "type": "object",
      "properties": {
        "Serial_Number": {
          "type": "string"
        },
        "Price": {
          "type": "string"
        },
        "Warehouse": {
          "type": "string"
        },
        "Remark_Inv": {
          "type": "string"
        },
        "Fabiao_Number": {
          "type": "string"
        },
        "Form_Number": {
          "type": "string"
        }
      }
    },
    "InvalidResponse": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}